BVAL-600 Only use the CustomRoleBlockProcessor for the DocBook output
